[html5] bugfix image

+++ html5/render/vue/utils/func.js
@@ -2,6 +2,7 @@
  * Mix properties into target object.
  */
 export function extend (to, from) {
+  if (!from) { return to }
   for (const key in from) {
     to[key] = from[key]
   }
@@ -12,6 +13,7 @@ export function extend (to, from) {
  * Mix specified properties into target object.
  */
 export function extendKeys (to, from, keys) {
+  if (!from) { return to }
   (keys || []).forEach(key => {
     to[key] = from[key]
   })
@@ -22,6 +24,7 @@ export function extendKeys (to, from, keys) {
  * Extract specified properties from src to target object.
  */
 export function extractKeys (to, from, keys) {
+  if (!from) { return to }
   (keys || []).forEach(key => {
     to[key] = from[key]
     delete from[key]
